departments/procurement/scraping_skill/requirements.txt
playwright
playwright-stealth
beautifulsoup4
pydantic
pandas
aiohttp
nest_asyncio
google-generativeai
